Enemy AI Improvement + Jumping
I added a generator of NavMesh links that I can manually change the settings of. This allowed the enemy to jump where he should (mostly at least). Added official unity script that changes the jumps into a curve. Also made it so enemy stops when he is close to you and prepared the script for attacking.